:recycle: #155 -- Make Checkout component not depend on cart store anymore
